Add Premailer support to automatically inline CSS
Premailer is a common way of automatically inline CSS styles. This change checks if the premailer-rails gem is available, and then calls the same hook that runs when an email is sent for real. By integrating this here, we ensure that the preview markup is identical to that which will be delivered.